Adding one recycler view in tasks (#23)

+++ app/src/main/java/com/example/pam_project/DatabaseHelper.java
@@ -83,7 +83,7 @@ public class DatabaseHelper {
 
     public static long getRandomFromArray(long[] array) {
         int rnd = new Random().nextInt(array.length);
-        return (long) array[rnd];
+        return array[rnd];
     }
 
     public static int getRandom(int min, int max) {
